lock on htft and profit loss fix

+++ src/components/gameDetails/bookmaker/index.tsx
@@ -110,7 +110,7 @@ const Bookmaker = ({ title, box, data, detail }) => {
             <div className="d-flex flex-row justify-content-between w-100">
               <span
                 className={`${
-                  parseInt(
+                  parseFloat(
                     detail?.profitLossDataMatch?.[
                       profitLossDataForMatchConstants[data?.type]?.A +
                         "_" +
@@ -139,7 +139,7 @@ const Bookmaker = ({ title, box, data, detail }) => {
                         detail?.id
                     ] === "0"
                     ? ""
-                    : parseInt(
+                    : parseFloat(
                         detail?.profitLossDataMatch?.[
                           profitLossDataForMatchConstants[data?.type]?.A +
                             "_" +
@@ -264,7 +264,7 @@ const Bookmaker = ({ title, box, data, detail }) => {
             <div className="d-flex flex-row justify-content-between w-100">
               <span
                 className={`${
-                  parseInt(
+                  parseFloat(
                     detail?.profitLossDataMatch?.[
                       profitLossDataForMatchConstants[data?.type]?.B +
                         "_" +
@@ -293,7 +293,7 @@ const Bookmaker = ({ title, box, data, detail }) => {
                         detail?.id
                     ] === "0"
                     ? ""
-                    : parseInt(
+                    : parseFloat(
                         detail?.profitLossDataMatch?.[
                           profitLossDataForMatchConstants[data?.type]?.B +
                             "_" +
@@ -419,7 +419,7 @@ const Bookmaker = ({ title, box, data, detail }) => {
               <div className="d-flex flex-row justify-content-between w-100">
                 <span
                   className={`${
-                    parseInt(
+                    parseFloat(
                       detail?.profitLossDataMatch?.[
                         profitLossDataForMatchConstants[data?.type]?.C +
                           "_" +
@@ -448,7 +448,7 @@ const Bookmaker = ({ title, box, data, detail }) => {
                           detail?.id
                       ] === "0"
                       ? ""
-                      : parseInt(
+                      : parseFloat(
                           detail?.profitLossDataMatch?.[
                             profitLossDataForMatchConstants[data?.type]?.C +
                               "_" +
